+++ src/apis/nested-comment/nested-comment.resolver.ts
@@ -1,7 +1,7 @@
 import { UseGuards } from '@nestjs/common';
 import { Args, Context, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
-import { NestedCommentEntity } from '@src/apis/nested_comment/entities/nested_comment.entity';
-import { NestedCommentService } from '@src/apis/nested_comment/nested_comment.service';
+import { NestedCommentEntity } from '@src/apis/nested-comment/entities/nested-comment.entity';
+import { NestedCommentService } from '@src/apis/nested-comment/nested-comment.service';
 import { GqlAuthAccessGuard } from '@src/commons/auth/gql-auth.guard';
 import { IContext } from '@src/commons/types/context';
