+++ src/apis/pick/picks.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { BoardEntity } from '@src/apis/board/entities/board.entity';
-import { FictionBoardEntity } from '@src/apis/fiction_board/entities/fiction_board.entity';
+import { FictionBoardEntity } from '@src/apis/fiction-board/entities/fiction-board.entity';
 import { PickEntity } from '@src/apis/pick/entities/pick.entity';
 import { UserEntity } from '@src/apis/user/entities/user.entity';
 import { Repository } from 'typeorm';
