+++ src/apis/user/user.module.ts
@@ -3,7 +3,7 @@ import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { ImageEntity } from '@src/apis/image/entities/image.entity';
 import { BoardEntity } from '@src/apis/board/entities/board.entity';
-import { FollowCountEntity } from '@src/apis/followCounts/followCount.entity';
+import { FollowCountEntity } from '@src/apis/follow-counts/follow-count.entity';
 import { MailService } from '@src/apis/mail/mail.service';
 import { UserEntity } from '@src/apis/user/entities/user.entity';
 import { UserResolver } from '@src/apis/user/user.resolver';
