+++ src/app.module.ts
@@ -3,14 +3,14 @@ import { CacheModule } from '@nestjs/cache-manager';
 import { Module } from '@nestjs/common';
 import { GraphQLModule } from '@nestjs/graphql';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AttendListModule } from '@src/apis/attend_list/attend_list.module';
+import { AttendListModule } from '@src/apis/attend-list/attend-list.module';
 import { AuthModule } from '@src/apis/auth/auth.module';
 import { BoardModule } from '@src/apis/board/board.module';
 import { CommentModule } from '@src/apis/comment/comment.module';
-import { FictionBoardModule } from '@src/apis/fiction_board/fiction_board.module';
+import { FictionBoardModule } from '@src/apis/fiction-board/fiction-board.module';
 import { FollowModule } from '@src/apis/follow/follow.module';
 import { EmailModule } from '@src/apis/mail/mail.module';
-import { NestedCommentModule } from '@src/apis/nested_comment/nested_comment.module';
+import { NestedCommentModule } from '@src/apis/nested-comment/nested-comment.module';
 import { PickModule } from '@src/apis/pick/picks.module';
 import { TagModule } from '@src/apis/tag/tag.module';
 import { UserModule } from '@src/apis/user/user.module';
